requirements.txt
tokenizers==0.12.1
scikit-learn==1.0.2
datasets==2.1.0
ipykernel==6.13.0
torch==1.11.0
transformers==4.18.0 #Needs rust compiler
google-cloud-aiplatform==1.11.0
kfp==1.8.12
black==22.3.0
google-cloud-pipeline-components==1.0.5
cloudml-hypertune==0.1.0.dev6